+++ src/applicationContext.xml
@@ -12,7 +12,7 @@
         <property name="username" value="root"/>
         <property name="password" value="rkdudmysql4_"/>
     </bean>
-    <bean id="userDao" class="springbook.user.dao.UserDao">
+    <bean id="userDao" class="springbook.user.dao.UserDaoJdbc">
         <property name="dataSource" ref="dataSource"/> <!--아직 dao의 get()메서드가 dataSource를 사용하므로 지우지 않음-->
     </bean>
 </beans>
