+++ src/test-applicationContext.xml
@@ -13,7 +13,7 @@
         <property name="username" value="root"/>
         <property name="password" value="rkdudmysql4_"/>
     </bean>
-    <bean id="userDao" class="springbook.user.dao.UserDao">
+    <bean id="userDao" class="springbook.user.dao.UserDaoJdbc">
         <property name="dataSource" ref="dataSource"/>
     </bean>
 </beans>
